# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kiroulou API Documentation",
        "description": "Documentation de l'API utilisée par l'application mobile KiRoulOu, faite avec Laravel",
        "contact": {
            "email": "bezedache29@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "https://localhost/api",
            "description": "Server API KiRoulOu"
        }
    ],
    "paths": {
...
        "/clubs": {
            "post": {
                "tags": [
                    "Clubs"
                ],
                "summary": "Add Club",
                "description": "Création d'un club",
                "operationId": "14ad7ea532d198f8fd2215e0da9511f1",
                "requestBody": {
                    "$ref": "#/components/requestBodies/AddClub"
                },
                "responses": {
                    "201": {
                        "description": "Club créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "club created"
                                        },
                                        "user_update": {
                                            "properties": {
                                                "is_club_admin": {
                                                    "description": "Le user devient admin du club qu'il vient de créer",
                                                    "type": "boolean"
                                                },
                                                "club_id": {
                                                    "description": "Le user est membre du club qu'il vient de créer",
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object",
                                            "example": {
                                                "is_club_admin": true,
                                                "club_id": 1
                                            }
                                        },
                                        "club": {
                                            "description": "Détails du club",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Club"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Déjà dans un club",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Vous faites déjà partie d'un club !"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/clubs?page={page}": {
            "get": {
                "tags": [
                    "Clubs"
                ],
                "summary": "All clubs",
                "description": "Tous les clubs",
                "operationId": "5c01cb71e1a1ced0e2acc1a9ed33a6f0",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ClubWithCounts"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/clubs/{club_id}/clubInformations": {
            "get": {
                "tags": [
                    "Clubs"
                ],
                "summary": "Club's informations",
                "description": "Les informations du club",
                "operationId": "63c1b3db67056d9071c41b4509f3fe3f",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/club_id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ClubInformations"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/clubs/{club_id}/followOrUnfollow": {
            "post": {
                "tags": [
                    "Clubs"
                ],
                "summary": "Follow/UnFollow club",
                "description": "Follow/UnFollow un club",
                "operationId": "af69a344716587613075fbb0f84ab25b",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/club_id"
                    }
                ],
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/Created"
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntity"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/clubs/{club_id}/requestToJoin": {
            "post": {
                "tags": [
                    "Clubs"
                ],
                "summary": "User membership request",
                "description": "Demande d'adhésion au club",
                "operationId": "848adbaaf99fd7b610ed94a7e19e70a1",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/club_id"
                    }
                ],
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/Created"
                    },
                    "403": {
                        "description": "Demande déjà en cours de traitement",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Votre demande est déjà en traitement pour ce club"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/clubs/{club_id}/acceptRequestToJoin": {
            "post": {
                "tags": [
                    "Clubs"
                ],
                "summary": "Accept the user's membership request",
                "description": "Acceptation de la demande d'adhésion d'un user au club",
                "operationId": "0872b4c80819300ae0644ed9745ab75a",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/club_id"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Accept la demande d'adhésion du user dans le club",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Accept Membership Request"
                                        },
                                        "club": {
                                            "$ref": "#/components/schemas/Club"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "$ref": "#/components/responses/Conflit"
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntity"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/clubs/{club_id}/changeAdmin": {
            "post": {
                "tags": [
                    "Clubs"
                ],
                "summary": "Change club admin",
                "description": "Change d'admin pour le club",
                "operationId": "b76c1afd3a67284d794456696c7b6769",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/club_id"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/ChangeAdmin"
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/Created"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ClubInformations": {
                "description": "Informations du club pour la scene informations du profil",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Club"
                    }
                ]
            },
...
        },
...
        "parameters": {
...
            "club_id": {
                "name": "club_id",
                "in": "path",
                "description": "ID du club",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
...
            "page": {
                "name": "page",
                "in": "path",
                "description": "Numéro de la page",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
...
        },
...
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "description": "S'authentifier avec email & password pour récupérer un token d'authentification",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}